Remove old codegen path. Refactor files in current codegen path for readability.
