+++ forc/src/ops/forc_deploy.rs
@@ -24,7 +24,6 @@ pub async fn deploy(command: DeployCommand) -> Result<fuel_tx::ContractId> {
 
     let DeployCommand {
         path,
-        use_orig_asm,
         print_finalized_asm,
         print_intermediate_asm,
         print_ir,
@@ -40,7 +39,6 @@ pub async fn deploy(command: DeployCommand) -> Result<fuel_tx::ContractId> {
 
     let build_command = BuildCommand {
         path,
-        use_orig_asm,
         print_finalized_asm,
         print_intermediate_asm,
         print_ir,
