+++ forc/src/ops/forc_run.rs
@@ -27,7 +27,6 @@ pub async fn run(command: RunCommand) -> Result<Vec<fuel_tx::Receipt>> {
 
     let build_command = BuildCommand {
         path: command.path,
-        use_orig_asm: command.use_orig_asm,
         print_finalized_asm: command.print_finalized_asm,
         print_intermediate_asm: command.print_intermediate_asm,
         print_ir: command.print_ir,
