+++ sway-core/src/semantic_analysis/ast_node/expression/match_expression/typed/typed_match_branch.rs
@@ -144,7 +144,6 @@ impl TypedMatchBranch {
         let new_result = TypedExpression {
             expression: TypedExpressionVariant::CodeBlock(TypedCodeBlock {
                 contents: code_block_contents,
-                whole_block_span: typed_result_span.clone(),
             }),
             return_type: typed_result.return_type,
             is_constant: IsConstant::No,
