+++ sway-core/src/type_engine/type_info.rs
@@ -479,96 +479,6 @@ impl TypeInfo {
         ok(name, vec![], vec![])
     }
 
-    pub(crate) fn size_in_bytes(&self, err_span: &Span) -> Result<u64, CompileError> {
-        Ok(self.size_in_words(err_span)? * 8)
-    }
-
-    /// Calculates the stack size of this type in words,
-    /// to be used when allocating stack memory for it.
-    pub(crate) fn size_in_words(&self, err_span: &Span) -> Result<u64, CompileError> {
-        match self {
-            // Each char is a byte, so the size is the num of characters / 8
-            // rounded up to the nearest word
-            TypeInfo::Str(len) => Ok((len + 7) / 8),
-            // Since things are unpacked, all unsigned integers are 64 bits.....for now
-            TypeInfo::UnsignedInteger(_) | TypeInfo::Numeric => Ok(1),
-            TypeInfo::Boolean => Ok(1),
-            TypeInfo::Tuple(fields) => Ok(fields
-                .iter()
-                .map(|field_type| {
-                    resolve_type(field_type.type_id, err_span)
-                        .expect("should be unreachable?")
-                        .size_in_words(err_span)
-                })
-                .collect::<Result<Vec<u64>, _>>()?
-                .iter()
-                .sum()),
-            TypeInfo::Byte => Ok(1),
-            TypeInfo::B256 => Ok(4),
-            TypeInfo::Enum { variant_types, .. } => {
-                // the size of an enum is one word (for the tag) plus the maximum size
-                // of any individual variant
-                Ok(1 + variant_types
-                    .iter()
-                    .map(|x| -> Result<_, _> { look_up_type_id(x.r#type).size_in_words(err_span) })
-                    .collect::<Result<Vec<u64>, _>>()?
-                    .into_iter()
-                    .max()
-                    .unwrap_or(0))
-            }
-            TypeInfo::Struct { fields, .. } => Ok(fields
-                .iter()
-                .map(|field| -> Result<_, _> {
-                    resolve_type(field.r#type, err_span)
-                        .expect("should be unreachable?")
-                        .size_in_words(err_span)
-                })
-                .collect::<Result<Vec<u64>, _>>()?
-                .iter()
-                .sum()),
-            // `ContractCaller` types are unsized and used only in the type system for
-            // calling methods
-            TypeInfo::ContractCaller { .. } => Ok(0),
-            TypeInfo::Contract => unreachable!("contract types are never instantiated"),
-            TypeInfo::ErrorRecovery => unreachable!(),
-            TypeInfo::Unknown
-            | TypeInfo::Custom { .. }
-            | TypeInfo::SelfType
-            | TypeInfo::UnknownGeneric { .. } => Err(CompileError::UnableToInferGeneric {
-                ty: self.friendly_type_str(),
-                span: err_span.clone(),
-            }),
-            TypeInfo::Ref(id, _sp) => look_up_type_id(*id).size_in_words(err_span),
-            TypeInfo::Array(elem_ty, count) => {
-                Ok(look_up_type_id(*elem_ty).size_in_words(err_span)? * *count as u64)
-            }
-            TypeInfo::Storage { .. } => Ok(0),
-        }
-    }
-
-    pub(crate) fn is_copy_type(&self, err_span: &Span) -> Result<bool, CompileError> {
-        match self {
-            // Copy types.
-            TypeInfo::UnsignedInteger(_)
-            | TypeInfo::Numeric
-            | TypeInfo::Boolean
-            | TypeInfo::Byte => Ok(true),
-            TypeInfo::Tuple(_) if self.is_unit() => Ok(true),
-
-            // Unknown types.
-            TypeInfo::Unknown
-            | TypeInfo::Custom { .. }
-            | TypeInfo::SelfType
-            | TypeInfo::UnknownGeneric { .. } => Err(CompileError::UnableToInferGeneric {
-                ty: self.friendly_type_str(),
-                span: err_span.clone(),
-            }),
-
-            // Otherwise default to non-copy.
-            _otherwise => Ok(false),
-        }
-    }
-
     pub fn is_uninhabited(&self) -> bool {
         match self {
             TypeInfo::Enum { variant_types, .. } => variant_types
